Renamed files, added new api directory in tests folder, reformatted books model test to conform to pylint style guide
