--- tRecorderApi/api/tests/test_api_chapters.py
@@ -1,54 +1,0 @@
-from django.test import TestCase
-from rest_framework import status
-from rest_framework.test import APIClient
-import random
-import string
-from ..models import Language, Anthology, Book, Version, Mode, Project, Chapter
-
-
-class ChapterApiTest(TestCase):
-
-    def setUp(self):
-        self.client = APIClient()
-        self.lang = Language.objects.create(slug='en-x-demo', name='english')
-        self.anthology = Anthology.objects.create(slug='ot', name="old testament")
-        self.book = Book.objects.create(name='mark', number=5, slug='mrk', anthology=self.anthology)
-        self.version = Version.objects.create(slug='ulb', name="Unlocked literal bible")
-        self.mode = Mode.objects.create(slug="chk", name="chunk", unit=1)
-        self.project = Project.objects.create(version=self.version, mode=self.mode,
-                                              anthology=self.anthology, language=self.lang,
-                                              book=self.book)
-        self.chap = Chapter.objects.create(number=1, checked_level=1, published=False, project=self.project)
-        self.chap2 = Chapter.objects.create(number=2, checked_level=2, published=False, project=self.project)
-        self.random_url = ''.join(random.choices(string.ascii_uppercase +
-                                                 string.digits,
-                                                 k=random.randint(1,15)))
-
-    def test_number_of_items_are_equal(self):
-        chapter_num = Chapter.objects.count()
-        response = self.client.get('/api/chapters/')
-        self.assertEqual(len(response.data), chapter_num)
-
-    def test_get_request_has_200_status_code(self):
-        response = self.client.get('/api/chapters/')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_get_request_with_id_has_200_status_code(self):
-        response = self.client.get('/api/chapters/1/')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_get_request_with_id_has_200_status_code(self):
-        response = self.client.get('/api/chapters/?id=1')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_get_request_with_non_existent_id_has_404_status_code(self):
-        response = self.client.get('/api/chapters/4/')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_random_text_as_parameter_gives_400_status_code(self):
-        response = self.client.get('/api/chapters/?'+self.random_url)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_response_contains_project(self):
-        response = self.client.get('/api/chapters/')
-        self.assertContains(response.data, self.project)
